Move scene impl. from pathos/render to pathos/scene
