+++ projects/PathosEngine/src/pathos/light/shadow_directional.h
@@ -2,7 +2,7 @@
 
 #include "gl_core.h"
 #include "pathos/camera/camera.h"
-#include "pathos/render/scene.h"
+#include "pathos/scene/scene.h"
 
 #include <vector>
