+++ projects/PathosEngine/src/pathos/render/deferred/deferredpass_pack_base.h
@@ -3,7 +3,7 @@
 //#include "pathos/mesh/mesh.h"
 #include "pathos/mesh/geometry.h"
 #include "pathos/material/material.h"
-#include "pathos/render/scene.h"
+#include "pathos/scene/scene.h"
 #include "pathos/camera/camera.h"
 #include "pathos/shader/uniform_buffer.h"
