+++ projects/PathosEngine/src/pathos/render/forward/renderpass_base.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "pathos/mesh/mesh.h"
-#include "pathos/render/scene.h"
+#include "pathos/scene/scene.h"
 #include "pathos/camera/camera.h"
 #include "pathos/light/shadow.h"
