+++ projects/PathosEngine/src/pathos/scene/scene.cpp
