+++ projects/PathosEngine/src/pathos/scene/scene.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "pathos/light/light.h"
+
 #include "glm/glm.hpp"
 #include <vector>
